+++ src/combat/lanes/LanesCreator.cs
@@ -3,13 +3,13 @@ using Godot.Collections;
 
 public class LanesCreator : Node2D
 {
-    Dictionary<int, double> yPositionForNumLanes = new Dictionary<int, double>()
+    Dictionary<int, float> yPositionForNumLanes = new Dictionary<int, float>()
     {
-        {1, 532},
-        {2, 327},
-        {3, 258},
-        {4, 224},
-        {5, 203}
+        {1, 532f},
+        {2, 327f},
+        {3, 258f},
+        {4, 224f},
+        {5, 203f}
     };
 
     // programmatically makes lanes at startup based on a config for how many lanes the specific conquest needs
@@ -20,17 +20,17 @@ public class LanesCreator : Node2D
         // set number of lanes and position of the parent of the lanes (this node) based off dict
         // hard coded, but no other way
         int numLanes = laneTypes.Count;
-        this.Position = new Vector2((float)700.5, (float)yPositionForNumLanes[numLanes]);
+        this.Position = new Vector2(700.5f, yPositionForNumLanes[numLanes]);
 
         // Calculate how much x and y space we have in total
         ColorRect TopBar = (ColorRect)GetParent().FindNode("TopBar");
         ColorRect BottomBar = (ColorRect)GetParent().FindNode("BottomBar");
         ColorRect ArmyBase = (ColorRect)GetParent().FindNode("ArmyBase");
-        double lanesXAvailable = GetViewportRect().Size.x - ArmyBase.RectSize.x;
-        double lanesYAvailable = GetViewportRect().Size.y - TopBar.RectSize.y - BottomBar.RectSize.y;
+        float lanesXAvailable = GetViewportRect().Size.x - ArmyBase.RectSize.x;
+        float lanesYAvailable = GetViewportRect().Size.y - TopBar.RectSize.y - BottomBar.RectSize.y;
 
         // Figure out how much y space for each alne
-        double yPixelsForEachLane = lanesYAvailable / numLanes;
+        float yPixelsForEachLane = lanesYAvailable / numLanes;
 
         // info about which lane in the list we're on -- used to set position of the lane later
         int laneIndex = 0;
@@ -43,12 +43,12 @@ public class LanesCreator : Node2D
             Lane newLane = GD.Load<PackedScene>("res://src/combat/lanes/Lane.tscn").Instance<Lane>();
             newLane.laneImg = laneDetails.image;
             newLane.Curve = laneDetails.curve;
-            newLane.Position = new Vector2(0, (float)yPixelsForEachLane * laneIndex);
+            newLane.Position = new Vector2(0, yPixelsForEachLane * laneIndex);
 
             // calculate scale of the lane texture
-            double imgXScale = lanesXAvailable / laneDetails.image.GetWidth();
-            double imgYScale = yPixelsForEachLane / laneDetails.image.GetHeight();
-            Vector2 newScaleVector = new Vector2((float)imgXScale, (float)imgYScale);
+            float imgXScale = lanesXAvailable / laneDetails.image.GetWidth();
+            float imgYScale = yPixelsForEachLane / laneDetails.image.GetHeight();
+            Vector2 newScaleVector = new Vector2(imgXScale, imgYScale);
 
             // get sprite and set scale
             Sprite newLaneSprite = newLane.GetNode<Sprite>("Sprite");
